requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
Jinja2==3.1.4
SQLAlchemy==2.0.36
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
PyMySQL==1.1.1
itsdangerous==2.2.0
